refactor: use camelCase
