+++ internal/handlers/user_handler_login.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"github.com/PitiNarak/condormhub-backend/internal/handlers/dto"
 	"github.com/PitiNarak/condormhub-backend/pkg/errorHandler"
-	"github.com/PitiNarak/condormhub-backend/pkg/http_response"
+	"github.com/PitiNarak/condormhub-backend/pkg/httpResponse"
 	"github.com/go-playground/validator"
 	"github.com/gofiber/fiber/v2"
 )
@@ -42,5 +42,5 @@ func (h *UserHandler) Login(c *fiber.Ctx) error {
 		UserInformation: *user,
 	}
 
-	return c.Status(fiber.StatusOK).JSON(http_response.SuccessResponse("Login successful", response))
+	return c.Status(fiber.StatusOK).JSON(httpResponse.SuccessResponse("Login successful", response))
 }
