+++ internal/handlers/user_handler_register.go
@@ -4,7 +4,7 @@ import (
 	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
 	"github.com/PitiNarak/condormhub-backend/internal/handlers/dto"
 	"github.com/PitiNarak/condormhub-backend/pkg/errorHandler"
-	"github.com/PitiNarak/condormhub-backend/pkg/http_response"
+	"github.com/PitiNarak/condormhub-backend/pkg/httpResponse"
 	"github.com/go-playground/validator"
 	"github.com/gofiber/fiber/v2"
 )
@@ -48,6 +48,6 @@ func (h *UserHandler) Register(c *fiber.Ctx) error {
 		UserInformation: *gormUser,
 	}
 
-	return c.Status(fiber.StatusCreated).JSON(http_response.SuccessResponse("user successfully registered", response))
+	return c.Status(fiber.StatusCreated).JSON(httpResponse.SuccessResponse("user successfully registered", response))
 
 }
