+++ internal/handlers/user_handler_reset_password.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/PitiNarak/condormhub-backend/internal/handlers/dto"
 	"github.com/PitiNarak/condormhub-backend/pkg/errorHandler"
-	"github.com/PitiNarak/condormhub-backend/pkg/http_response"
+	"github.com/PitiNarak/condormhub-backend/pkg/httpResponse"
 	"github.com/go-playground/validator"
 	"github.com/gofiber/fiber/v2"
 )
@@ -42,7 +42,7 @@ func (h *UserHandler) ResetPassword(c *fiber.Ctx) error {
 	if err != nil {
 		return err
 	}
-	return c.Status(fiber.StatusOK).JSON(http_response.SuccessResponse("password reset successfully", fiber.Map{
+	return c.Status(fiber.StatusOK).JSON(httpResponse.SuccessResponse("password reset successfully", fiber.Map{
 		"userInformation": user,
 		"accessToken":     tokenString,
 	}))
